Link search results to passages
